Fixed stuff - see description
Fixed player rotation in the rig/model a lot of the spine and head was rotated randomly so I reverted those to 0 again.

Fixed bugs with some of the animation and attempted to refix the link that was broken to reset the animation. I have left a couple notes in the code at where the problem seems to lie, but I haven't quite found it yet. It's not letting it come out of the animation if you don't let go of the fire button, which shouldn't be the case.